pyproject.toml
[project]
name = "uv-fastapi-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.112.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "supabase>=2.11.0",
    "workos>=5.10.0",
]
